+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Pass in request headers instead of the actual request to RequestedVariant;
 * Slit testing framework helpers from acceptance test frameworks. This means we
   can now use a combination of RSpec/Minitest and Capybara/ActiveSupport test
   cases.